Add files via upload
The above file contains the logical explanation of the sorting techniques in the form of the code. Insertion sort, selection sort are the major concepts ion the data structures and algorithms where the people uses the key element for the searching techniques.These can be helpful for understanding the concepts.